better way to find the actual expression
In gomega expression, the current code knows how to skip the helper
method like `WithOffset()` and get to the actual method (e.g. `Expect`).

The problem is that this logic is done by comparing to a static list of
method names. If a new method will be added, we'll have to add it to the
list each time.

This commit fixes the issue by skiping any non-actual method.

Also, move things around for better oerdering.
